third_party/WebKit/LayoutTests/dom/node/missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
var documentFragment = document.createDocumentFragment();
var documentType = document.doctype;
var element = document.documentElement;
var characterData = document.createTextNode("");
[
    "compareDocumentPosition()",
    "contains()",
    "isDefaultNamespace()",
    "isEqualNode()",
    "lookupNamespaceURI()",
    "lookupPrefix()"
].forEach(function(expr)
{
    shouldThrow("document." + expr);
    shouldThrow("documentFragment." + expr);
    shouldThrow("documentType." + expr);
    shouldThrow("element." + expr);
    shouldThrow("characterData." + expr);
});
</script>
</body>
</html>
