third_party/blink/web_tests/dom/node/missing-arguments-expected.txt

PASS document.compareDocumentPosition() threw exception TypeError: Failed to execute 'compareDocumentPosition' on 'Node': 1 argument required, but only 0 present..
PASS documentFragment.compareDocumentPosition() threw exception TypeError: Failed to execute 'compareDocumentPosition' on 'Node': 1 argument required, but only 0 present..
PASS documentType.compareDocumentPosition() threw exception TypeError: Failed to execute 'compareDocumentPosition' on 'Node': 1 argument required, but only 0 present..
PASS element.compareDocumentPosition() threw exception TypeError: Failed to execute 'compareDocumentPosition' on 'Node': 1 argument required, but only 0 present..
PASS characterData.compareDocumentPosition() threw exception TypeError: Failed to execute 'compareDocumentPosition' on 'Node': 1 argument required, but only 0 present..
PASS document.contains() threw exception TypeError: Failed to execute 'contains' on 'Node': 1 argument required, but only 0 present..
PASS documentFragment.contains() threw exception TypeError: Failed to execute 'contains' on 'Node': 1 argument required, but only 0 present..
PASS documentType.contains() threw exception TypeError: Failed to execute 'contains' on 'Node': 1 argument required, but only 0 present..
PASS element.contains() threw exception TypeError: Failed to execute 'contains' on 'Node': 1 argument required, but only 0 present..
PASS characterData.contains() threw exception TypeError: Failed to execute 'contains' on 'Node': 1 argument required, but only 0 present..
PASS document.isDefaultNamespace() threw exception TypeError: Failed to execute 'isDefaultNamespace' on 'Node': 1 argument required, but only 0 present..
PASS documentFragment.isDefaultNamespace() threw exception TypeError: Failed to execute 'isDefaultNamespace' on 'Node': 1 argument required, but only 0 present..
PASS documentType.isDefaultNamespace() threw exception TypeError: Failed to execute 'isDefaultNamespace' on 'Node': 1 argument required, but only 0 present..
PASS element.isDefaultNamespace() threw exception TypeError: Failed to execute 'isDefaultNamespace' on 'Node': 1 argument required, but only 0 present..
PASS characterData.isDefaultNamespace() threw exception TypeError: Failed to execute 'isDefaultNamespace' on 'Node': 1 argument required, but only 0 present..
PASS document.isEqualNode() threw exception TypeError: Failed to execute 'isEqualNode' on 'Node': 1 argument required, but only 0 present..
PASS documentFragment.isEqualNode() threw exception TypeError: Failed to execute 'isEqualNode' on 'Node': 1 argument required, but only 0 present..
PASS documentType.isEqualNode() threw exception TypeError: Failed to execute 'isEqualNode' on 'Node': 1 argument required, but only 0 present..
PASS element.isEqualNode() threw exception TypeError: Failed to execute 'isEqualNode' on 'Node': 1 argument required, but only 0 present..
PASS characterData.isEqualNode() threw exception TypeError: Failed to execute 'isEqualNode' on 'Node': 1 argument required, but only 0 present..
PASS document.lookupNamespaceURI() threw exception TypeError: Failed to execute 'lookupNamespaceURI' on 'Node': 1 argument required, but only 0 present..
PASS documentFragment.lookupNamespaceURI() threw exception TypeError: Failed to execute 'lookupNamespaceURI' on 'Node': 1 argument required, but only 0 present..
PASS documentType.lookupNamespaceURI() threw exception TypeError: Failed to execute 'lookupNamespaceURI' on 'Node': 1 argument required, but only 0 present..
PASS element.lookupNamespaceURI() threw exception TypeError: Failed to execute 'lookupNamespaceURI' on 'Node': 1 argument required, but only 0 present..
PASS characterData.lookupNamespaceURI() threw exception TypeError: Failed to execute 'lookupNamespaceURI' on 'Node': 1 argument required, but only 0 present..
PASS document.lookupPrefix() threw exception TypeError: Failed to execute 'lookupPrefix' on 'Node': 1 argument required, but only 0 present..
PASS documentFragment.lookupPrefix() threw exception TypeError: Failed to execute 'lookupPrefix' on 'Node': 1 argument required, but only 0 present..
PASS documentType.lookupPrefix() threw exception TypeError: Failed to execute 'lookupPrefix' on 'Node': 1 argument required, but only 0 present..
PASS element.lookupPrefix() threw exception TypeError: Failed to execute 'lookupPrefix' on 'Node': 1 argument required, but only 0 present..
PASS characterData.lookupPrefix() threw exception TypeError: Failed to execute 'lookupPrefix' on 'Node': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
